Addition of new function to compute the QTL effect of single position puting the other QTLs as cofactor.
The reason to develop such a function was that QTL_effects_GE and QTL_effects_oneS did not work with to many QTL elements to compute.

Therefore I made QTL_effects_GE_1QTL and QTL_effects_oneS_1QTL.

It seems to work. I will now implement a function to compute the R2